+++ README.md
@@ -7,7 +7,6 @@ This is a boilerplate for the Husky Hunt App API. Please see [this](https://mher
 **API Team**: 
 - *Co-lead* Matt
 - *Co-lead* Dan
-- Michael
 - Ben
 
 **Web Team**:
@@ -21,3 +20,8 @@ This is a boilerplate for the Husky Hunt App API. Please see [this](https://mher
 - everyone else
 
 
+# Creating Routes
+
+At a high level, a REST API is organized into routes. Each route provides a logically grouped subset of the API functionality. For example, `www.api.com/map/...` might represent the `map` route, and provide all functionality related to adding or retrieving map information. 
+
+In the code, the implementations for different routes are similarly partitioned. In `controllers`, there is a file for each high level route, and these files are imported in `controllers/main.ts` and then added as routes in the express app instance.
